+++ src/generators/dom/visitors/Element.js
@@ -1,5 +1,5 @@
-import CodeBuilder from '../../utils/CodeBuilder.js';
-import deindent from '../../utils/deindent.js';
+import CodeBuilder from '../../../utils/CodeBuilder.js';
+import deindent from '../../../utils/deindent.js';
 import addElementAttributes from './attributes/addElementAttributes.js';
 import Component from './Component.js';
