+++ src/generators/dom/visitors/RawMustacheTag.js
@@ -1,4 +1,4 @@
-import deindent from '../../utils/deindent.js';
+import deindent from '../../../utils/deindent.js';
 
 export default {
 	enter ( generator, node ) {
