+++ src/generators/dom/visitors/attributes/lookup.js
