+++ src/generators/shared/utils/counter.js
