+++ src/generators/shared/utils/getGlobals.js
