+++ src/generators/shared/utils/getOutro.js
