+++ src/index.js
@@ -1,7 +1,7 @@
 import parse from './parse/index.js';
 import validate from './validate/index.js';
-import generate from './generate/index.js';
-import generateSSR from './server-side-rendering/compile.js';
+import generate from './generators/dom/index.js';
+import generateSSR from './generators/server-side-rendering/index.js';
 
 function normalizeOptions ( options ) {
 	return Object.assign( {
